Fix platform-dependent tests.
